requirements.txt
## data science libraries 
pandas
numpy
#scikit-learn
numpy 
keras

# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
attrs
pyrsistent
importlib-metadata
flask
cython
coverage
joblib
pyyaml

